requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
idna==3.4
python-decouple==3.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
